Reorganize music
This moves all the music files to a separate folder. It also turns the
music and sounds folder into Python packages using code from music.py
and sounds.py respectively. For now, the new __init__.py scripts depend
on base.py in the parent directory and uses somewhat of a hack to import
that script. In the future, that hack will be removed.

Closes #7
